requirements.txt
asgiref==3.2.7
attrs==19.3.0
awesome-slugify==1.6.5
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
factory-boy==3.0.1
Faker==4.1.2
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
regex==2020.7.14
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
Unidecode==0.4.21
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0

